type: chore
I changed where I am converting the type of the actions tensor in the experiences to when they are sampled instead of  right as they are gathered. I didn't think this should make a big difference because it is essentially moving a conversion from when the value is used to when it is first retrieved and there isn't anything done to it in the meantime. But that is the way it is in the udacity DQN code and at this point I'm just trying to achieve parity with that.

+++ reinforcement_learning/policy.py
@@ -50,7 +50,7 @@ def learn(learning_network, target_network, optimizer, experience_batch, gamma):
     y = rewards + (gamma * reshaped_target_action_value_max * (1 - dones))
 
     # getting the action values from the model that is learning
-    learning_action_value_estimates = learning_network(states).gather(1, actions.long())
+    learning_action_value_estimates = learning_network(states).gather(1, actions)
 
     # loss function, mean squared error
     loss = F.mse_loss(learning_action_value_estimates, reshaped_target_action_value_max)
